backend/ai_engine/requirements.txt
fastapi
uvicorn
langgraph
langchain-core
google-generativeai
paramiko
python-dotenv
pydantic
httpx
websockets
